fix log and example
